requirements.txt
fastapi
uvicorn[standard]
sse-starlette
langchain
langchain-anthropic
langgraph
pinecone-client
youtube-transcript-api
python-dotenv
anthropic
